Separate out tests into webdriver, renderer, and unit, to avoid use globbing which has CI/cross platform issues.
