+++ test/unit/lint/test-devdeps.js
@@ -4,7 +4,7 @@
 import expect from 'expect';
 import fs from 'fs-promise';
 import without from 'lodash/without';
-import { autoFailingAsyncTest } from '../utils/async';
+import { autoFailingAsyncTest } from '../../utils/async';
 import { REQUIRES_REGEX, IMPORTS_REGEX, globMany, regexFiles } from './shared.js';
 
 const all = '/**/';
